Change jar-FileName-class-seperator char from '!' to '|'(<-Pipe)
That was just for illustrating / inspiration
Well I've seen there are like 3 other locations in the code a global search for '! and "! will find them.

The main idea here was to use some char as separator that is not a invalid char for filenames and path and so is unique.

Well I checked out the thing about JarURLConnection like this:
**jar:file:/home/duke/duke.jar!/com/BintrayExtension.class**
The thing is that accessing files inside a JAR is only working if it's done via a JarURLConnection.
It's not applied to 'normal' File inputstreams. 
So I see no reason to hold on at '!' as seperator.